Mapbox integrated  with Streamlit using the pydeck library.
