+++ homeworks/zsoltcsengeri/8_streamlit/frontend.py
@@ -1,5 +1,6 @@
 import streamlit as st
-from openweather_api_call import fetch_weather
+from temp_humidity_wind_api_call import fetch_weather
+import pydeck as pdk
 
 # Set up the title of the app
 st.title("Robot Dreams Python - Weather Map & Data Visualization APP")
@@ -10,7 +11,7 @@ city = st.text_input("Enter the city name", "London")
 # Fetch and display weather data
 if city:
     data = fetch_weather(city)
-    
+
     # Check if the API call was successful
     if data.get("cod") == 200:
         # Extract weather information
@@ -26,7 +27,30 @@ if city:
         col1.metric("Temperature", temperature, f"Feels like {feels_like}")
         col2.metric("Humidity", humidity)
         col3.metric("Wind Speed", wind_speed)
-        
+
+        # Get latitude and longitude for the city
+        latitude = data["coord"]["lat"]
+        longitude = data["coord"]["lon"]
+
+        # Weather map display using pydeck
+        st.subheader("Weather Map")
+        layer = pdk.Layer(
+            "ScatterplotLayer",
+            data=[{"lat": latitude, "lon": longitude}],
+            get_position=["lon", "lat"],
+            get_color="[200, 30, 0, 160]",
+            get_radius=1000,
+        )
+        view_state = pdk.ViewState(
+            latitude=latitude,
+            longitude=longitude,
+            zoom=8,
+            pitch=0,
+        )
+
+        # Render map
+        st.pydeck_chart(pdk.Deck(layers=[layer], initial_view_state=view_state))
+
     else:
         # Display an error message if the API call fails
         st.error(data.get("message", "Error fetching data"))
